requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-generativeai
html2pdf
jinja2
python-multipart
beautifulsoup4